불필요한 TodoPage를 제거하고 그 내용을 App으로 옮겨라
