+++ src/App.jsx
@@ -1,7 +1,12 @@
-import TodoPage from './presentational/TodoPage';
+import InputContainer from './container/InputContainer';
+import ListContainer from './container/ListContainer';
 
 export default function App() {
   return (
-    <TodoPage />
+    <div>
+      <h1>To-do</h1>
+      <InputContainer />
+      <ListContainer />
+    </div>
   );
 }
