--- src/presentational/TodoPage.jsx
@@ -1,12 +1,0 @@
-import InputContainer from '../container/InputContainer';
-import ListContainer from '../container/ListContainer';
-
-export default function TodoPage() {
-  return (
-    <div>
-      <h1>To-do</h1>
-      <InputContainer />
-      <ListContainer />
-    </div>
-  );
-}
